generalize views even more
basically we are ready to retire the dual view and use the more generalized quad views
quad views should automatically resize as players connect and disconnect.

we do this by doing a really inefficient check for how many quad cols and rows are visible. then we use this info to figure out how big to make the view