delete obj_global
this caused html5 to fail. I think there's a race condition on when objects get loaded in the first room. if obj_display_manager gets run first, it runs the goto next room command and I think any other objects don't get properly run.

this means on windows, obj_global was successfully running (set_globals() from create event), but on html5, for some reason, it never runs. therefore all the global variables are undefined.

this was kinda bad when we moved the global.cell_size to set_globals(), which was when html5 initially broke, but as soon as we moved ALL the stuff to, now, init_globals(), nothing worked at all.

the solution is to run init_globals() in the create event of the first room, to ENSURE it gets run.  this also has the added benefit of eliminating an object